Remove locations from regions that require a level above the level cap
Many regions have a corresponding 'Access' item which is required to
gain access to that region. When the 'Access' item has a level
requirement above the level cap, it is removed from the item pool, but
this would cause the region, and all locations within it, to become
unreachable.

By finding these 'Access' items after region creation, their level
requirements can be checked against the level cap. If the requirement is
above the level cap, then all locations are now removed from the region.